matchbox/requirements.txt
PyYAML
openai
